requirements.txt
amqp==2.5.0
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
billiard==3.6.0.0
bleach==3.3.0
celery==4.3.0
certifi==2019.6.16
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.1
colorama==0.4.4
cryptography==3.4.7
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-timezone-field==3.0
djangorestframework==3.12.4
entrypoints==0.3
idna==2.8
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kombu==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
neo4j==4.3.3
neo4j-driver==4.3.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.2.0
pytz==2019.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.1.0
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.3.0
Send2Trash==1.7.1
six==1.12.0
sqlparse==0.3.0
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.25.3
vine==1.3.0
virtualenv==16.6.0
wcwidth==0.2.5
webencodings==0.5.1
